+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Version 21.1-SNAPSHOT
 
 ADDED:
 
-- new method that determines the index to insert to a given list with the binary search algorithm
+- new extension method in class ListExtensions that determines the index to insert from a given list with a given element and searches with the 'binary search' algorithm
 
 Version 21
 -------------